moves rest of API types
