+++ app/layout/projects/new/form/country-region-selector/types.ts
@@ -1,6 +1,6 @@
 import { MouseEvent } from 'react';
 
-import { Country, Region, SubRegion } from 'types/location';
+import { Country, Region, SubRegion } from 'types/api/location';
 
 export default interface CountryRegionSelectorProps {
   country?: Country['id'];
