+++ app/types/api/job.ts
